+++ packages/cli/core/modern.config.js
@@ -3,11 +3,4 @@ module.exports = {
   output: {
     disableSourceMap: true,
   },
-  testing: {
-    jest: {
-      collectCoverage: true,
-      collectCoverageFrom: ['src/**/*.ts'],
-      coveragePathIgnorePatterns: ['/node_modules/'],
-    },
-  },
 };