+++ packages/cli/plugin-ssr/node.d.ts
@@ -1,6 +1,0 @@
-// eslint-disable-next-line filenames/match-exported
-import plugin from './dist/types/index.node';
-
-export * from './dist/types/index.node';
-
-export default plugin;