+++ packages/cli/static-hosting/tests/index.test.ts
@@ -1,0 +1,7 @@
+import plugin from '../src';
+
+describe('plugin-static-hosting', () => {
+  it('default', () => {
+    expect(plugin).toBeDefined();
+  });
+});