+++ packages/toolkit/create/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/create
 
+## 1.2.4
+
+### Patch Changes
+
+- f73fee4b: feat: upgrade codesmith version to add get npm package timeout's time
+
 ## 1.2.2
 
 ### Patch Changes